refactor: move Validation interface to a folder
